converted block and chunk positions into discrete ivecs

+++ src/render/gl/gl-vao.cpp
@@ -16,7 +16,8 @@ void GLVertexArray::enable() {
 }
 
 ChunkVertexArray::ChunkVertexArray() :
-    vertices(std::make_unique<GLArrayBuffer<glm::vec3>>(0, 3)),
+    // packedVertices(std::make_unique<GLArrayBuffer<glm::uint32>>(0, 1)),
+    vertices(std::make_unique<GLArrayBuffer<glm::ivec3>>(0, 3)),
     normals(std::make_unique<GLArrayBuffer<glm::vec3>>(2, 3)),
     uvs(std::make_unique<GLArrayBuffer<glm::vec2>>(1, 2)),
     texIDs(std::make_unique<GLArrayBuffer<int>>(3, 1)),
@@ -26,6 +27,16 @@ ChunkVertexArray::ChunkVertexArray() :
 
 void ChunkVertexArray::writeToBuffers(const IndexedMeshData &data) const {
     glBindVertexArray(objectID);
+
+    // std::vector<glm::uint32> packedVerticesData;
+    //
+    // for (size_t i = 0; i < data.vertices.size(); i++) {
+    //     glm::uint32 packedVertex = 0;
+    //     packedVertex |= ...
+    // }
+    //
+    // packedVertices->write(packedVerticesData);
+
     vertices->write(data.vertices);
     uvs->write(data.uvs);
     normals->write(data.normals);
